% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort_freq.R
\name{abort_freq}
\alias{abort_freq}
\title{abort_freq}
\usage{
abort_freq(rel = FALSE)
}
\arguments{
\item{rel}{logical. If TRUE, it returns the relative frequency}
}
\value{
table
}
\description{
Computes the absolute or relative frequency of the view on abortion from the dataset 'canada_ab'
}
\examples{
\dontrun{
abort_freq(rel=TRUE)
}
}
