% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absrel_freq.R
\name{absrel_freq}
\alias{absrel_freq}
\title{absrel_freq}
\usage{
absrel_freq(x, rel = FALSE)
}
\arguments{
\item{x}{vector. For computing frequency}

\item{rel}{logical. If TRUE, returns the relative frequency}
}
\value{
table
}
\description{
Computes the absolute or relative frequency of elements of a vector
}
\examples{
\dontrun{
absrel_fre(x=abortion,rel=TRUE)
}

}
