% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{canada_ab}
\alias{canada_ab}
\title{canada_ab}
\format{dataframe with 2231 rows and 9 columns:
\describe{
\item{id}{numeric vector for Household ID number}
\item{province}{a factor with levels for province}
\item{population}{numeric vector for population over age 17 of the province}
\item{weights}{numeric vector for weight sample of the size of population}
\item{gender}{a factor with levels Male, Female}
\item{abortion}{factor with levels No, Yes answering the question 'should abortion be banned?'}
\item{importance}{factor for indicating the importance of religion}
\item{education}{factor for indicating the level of education}
\item{urban}{factor for indicating the residence}
}}
\source{
\url{https://vincentarelbundock.github.io/Rdatasets/datasets.html}
}
\usage{
canada_ab
}
\description{
A dataset containing data from the 2011 Canadian National Election Study regarding banning abortion
}
\keyword{datasets}
