% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{chicken_weight}
\alias{chicken_weight}
\title{chicken_weight}
\format{dataframe with 578 rows and 4 columns:
\describe{
\item{weight}{numeric vector with the weight of the chickens (gm)}
\item{Time}{numeric vector with the days since birth}
\item{Chick}{ordered factor that identifies the chickens. They are ordered according to similar diet and final weight}
\item{Diet}{a factor with levels 1,..,4 with the diet received}
}}
\source{
\url{https://vincentarelbundock.github.io/Rdatasets/datasets.html}
}
\usage{
chicken_weight
}
\description{
A dataset containing the effect of diet on early growth of chickens
}
\keyword{datasets}
