% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_dataset.R
\name{online_dataset}
\alias{online_dataset}
\title{online_dataset}
\usage{
online_dataset(URL = NULL)
}
\arguments{
\item{URL}{url from 'https://vincentarelbundock.github.io/Rdatasets/datasets.html'}
}
\value{
data frame
}
\description{
Gets online dataset in csv format from 'https://vincentarelbundock.github.io/Rdatasets/datasets.html'
}
\examples{
\dontrun {
online_dataset('https://vincentarelbundock.github.io/Rdatasets/csv/boot/amis.csv')
}

}
