% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ds.R
\name{plot_ds}
\alias{plot_ds}
\title{plot_ds}
\usage{
plot_ds(dataset, x, y = NULL, colour = NULL)
}
\arguments{
\item{dataset}{dataframe. Name of the dataset to plot}

\item{x}{numeric vector or vector of factors Name of the column}

\item{y}{numeric vector. Name of the column for scatterplot}

\item{colour}{numeric or vector of factors. Name of the column to paint the data in the plot}
}
\value{
scatterplot if numeric vector, barplot if vector of factors
}
\description{
It plots the data depending on the data used
}
\examples{
\dontrun{
plot_ds(dataset=canada_ab,x=canada_ab$abortion,colour=canada_ab$importance)
}

}
