% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_signif.R
\name{regr_signif}
\alias{regr_signif}
\title{regr_signif}
\usage{
regr_signif(y, x)
}
\arguments{
\item{y}{numeric vector. Dependent variable}

\item{x}{numeric vector. Independent variable}
}
\value{
table, str
}
\description{
Regress a variable y on a variable x
If the intercept value is non-significant, it computes again the regression without the intercept. It also returns if the regressed variable is significant or not
}
\examples{
\dontrun{
regr_signif(y=chicken_weight$weight,x=chicken_weight$Time)
}

}
