% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condfreq_abort.R
\name{condfreq_abort}
\alias{condfreq_abort}
\title{condfreq_abort}
\usage{
condfreq_abort(x)
}
\arguments{
\item{x}{characher. Name of the column of the dataset for computing the conditioned frequency}
}
\value{
list
}
\description{
Computes the frequency of the view on abortion from the dataset 'canada_ab',
conditioned to one of the other variables in the dataset.
}
\examples{
\dontrun{
confreq_abort(gender)
}
}
